Fix : Award ReFactoring
console.log 제거 및 에러 처리 구문 추가

해결

+++ back/src/routers/awardRouter.js
@@ -8,6 +8,12 @@ const awardRouter = Router();
 awardRouter.get("/awardlist/:user_id", login_required, async (req, res, next) => {
     try {
 
+        if (is.emptyObject(req.query)) {
+            throw new Error(
+                "페이지네이션을 위한 쿼리를 확인해주세요."
+            );
+        }
+
         const page = Number(req.query.page || 1);
         const perPage = Number(req.query.perPage || 3);
 
@@ -17,7 +23,6 @@ awardRouter.get("/awardlist/:user_id", login_required, async (req, res, next) =>
             perPage : perPage,
         }
 
-        console.log("router");
 
         const award = await awardService.getAwards({ getAwards });
 
@@ -35,11 +40,15 @@ awardRouter.get("/awardlist/:user_id", login_required, async (req, res, next) =>
 awardRouter.get("/awards/:id", login_required, async (req, res, next) => {
     try {
 
+        if (is.emptyObject(req.params)) {
+            throw new Error(
+                "award id가 없습니다. 다시 확인해주세요."
+            );
+        }
+
         const getAward = {
             id : req.params.id,
         }
-
-        console.log("router");
 
         const award = await awardService.getAward({ getAward });
 
@@ -55,15 +64,19 @@ awardRouter.get("/awards/:id", login_required, async (req, res, next) => {
     }
 })
 awardRouter.post("/award/create", login_required, async (req, res, next) => {
-    
     try {
+
+        if (is.emptyObject(req.body)) {
+            throw new Error(
+                "데이터를 다시 확인해주세요."
+            );
+        }
+
         const newAward = {
             user_id : req.body.user_id,
             award : req.body.award,
             description : req.body.description,
         }
-
-        console.log("router");
 
         const award = await awardService.addAward({ newAward });
 
@@ -81,13 +94,19 @@ awardRouter.post("/award/create", login_required, async (req, res, next) => {
 
 awardRouter.delete("/awards/:id", login_required, async (req, res, next) => {
     try {
+
+        if (is.emptyObject(req.params)) {
+            throw new Error(
+                "award id가 없습니다. 다시 확인해주세요."
+            );
+        }
+
         const user_id = req.currentUserId;
         const deleteAward = {
             user_id : user_id,
             id : req.params.id,
         }
 
-        console.log("router");
         const award = await awardService.deleteAward({ deleteAward });
 
         if (award.errorMessage) {
@@ -104,14 +123,17 @@ awardRouter.delete("/awards/:id", login_required, async (req, res, next) => {
 
 awardRouter.post("/award/:id", login_required, async (req, res, next) => {
     try {
-        //const user_id = req.currentUserId;
+        if (is.emptyObject(req.params) || is.emptyObject(req.body)) {
+            throw new Error(
+                "award 수정에 실패했습니다. 다시 확인해주세요."
+            );
+        }
+
         const updateAward = {
             id : req.params.id,
             changeAward : req.body.changeAward,
             changeDescription : req.body.changeDescription
         }
-
-        console.log("router");
 
         const award = await awardService.updateAward({ updateAward });
